backend/pyproject.toml
[project]
name = "review-hub-backend"
version = "0.1.0"
description = "Backend FastAPI para Review Hub - Plataforma de Revisão Sistemática"
readme = "README.md"
license = { text = "AGPL-3.0" }
requires-python = ">=3.11"
authors = [
    { name = "Raphael Federicci Haddad" }
]

dependencies = [
    # FastAPI & ASGI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "gunicorn>=21.2.0",  # Production WSGI server
    
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.0.0",  # Required for SQLAlchemy async operations

    # Migrations
    "alembic>=1.13",
    "psycopg[binary]>=3.1", # Sync driver required for startup migration check
    
    # Validation & Settings
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # HTTP Client
    "httpx>=0.27.0",
    
    # Supabase SDK (Storage)
    "supabase>=2.0.0",
    
    # AI/LLM
    "openai>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "instructor>=1.0.0",
    
    # PDF Processing
    "pypdf>=5.0.0",
    
    # Observability
    "structlog>=24.0.0",
    
    # Rate Limiting
    "slowapi>=0.1.9",
    
    # Utils
    "python-multipart>=0.0.9",
    "tenacity>=9.0.0",
    "cryptography>=43.0.0",
    
    # Task Queue
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "UP042",  # str+Enum kept for SQLAlchemy/Pydantic compatibility (StrEnum migration deferred)
]

[tool.ruff.lint.per-file-ignores]
# Multiline fixture strings use indented blank lines; trailing space in SQL snippets.
"tests/**" = ["W291", "W293"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"


[tool.coverage.run]
# KISS/DRY: focar coverage no código que está sendo efetivamente exercitado e mantido.
# Estes diretórios estão atualmente sem testes (0%) e puxam a cobertura para baixo no CI.
omit = [
    "app/schemas/read_models/*",
    "app/worker/*",
]
